-Creados metodos para recibir imagen de nota segun id, por hash o por bytes

+++ grails-app/controllers/okylifeapi/OkiBarController.groovy
@@ -23,7 +23,7 @@ class OkiBarController {
         respond new OkiBar(params)
     }
 
-    def getOkiBarByActivity(int activityId) {
+    def getOkiBarByActivity(long activityId) {
         def activityInstance = Activity.findById(activityId)
         if (activityInstance) {
             def okiBar = activityInstance.getOkiBar()
